refactor: create @components alias
