+++ .vscode/typescriptreact.code-snippets
@@ -83,8 +83,8 @@
     "prefix": "np",
     "body": [
       "import * as React from 'react';\n",
-      "import Layout from '@/components/layout/Layout';",
-      "import Seo from '@/components/Seo';\n",
+      "import Layout from '@components/layout/Layout';",
+      "import Seo from '@components/Seo';\n",
       "export default function ${1:${TM_FILENAME_BASE/(^[a-zA-Z])(.*)/${1:/upcase}${2}/}}Page() {",
       "  return (",
       "    <Layout>",
