+++ src/components/layout/Header.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import UnstyledLink from '@/components/links/UnstyledLink';
+import UnstyledLink from '@components/links/UnstyledLink';
 
 const links = [
   { href: '/', label: 'Route 1' },
