+++ src/pages/404.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { RiAlarmWarningFill } from 'react-icons/ri';
 
-import Layout from '@/components/layout/Layout';
-import ArrowLink from '@/components/links/ArrowLink';
-import Seo from '@/components/Seo';
+import Layout from '@components/layout';
+import ArrowLink from '@components/links/ArrowLink';
+import Seo from '@components/Seo';
 
 export default function NotFoundPage() {
   return (
