+++ src/pages/_app.tsx
@@ -8,8 +8,8 @@ import '@/styles/globals.css';
 import '@/styles/colors.css';
 import 'react-toastify/dist/ReactToastify.min.css';
 
-import Layout from '@/components/layout/Layout';
-import Navbar from '@/components/navbar';
+import Layout from '@components/layout';
+import Navbar from '@components/navbar';
 
 import { isProd } from '@/constant/env';
 import { AuthProvider } from '@/contexts/authContext';
