+++ src/pages/components.tsx
@@ -10,18 +10,18 @@ import {
   HiPlus,
 } from 'react-icons/hi';
 
-import Button from '@/components/buttons/Button';
-import IconButton from '@/components/buttons/IconButton';
-import TextButton from '@/components/buttons/TextButton';
-import Layout from '@/components/layout/Layout';
-import ArrowLink from '@/components/links/ArrowLink';
-import ButtonLink from '@/components/links/ButtonLink';
-import PrimaryLink from '@/components/links/PrimaryLink';
-import UnderlineLink from '@/components/links/UnderlineLink';
-import UnstyledLink from '@/components/links/UnstyledLink';
-import NextImage from '@/components/NextImage';
-import Seo from '@/components/Seo';
-import Skeleton from '@/components/Skeleton';
+import Button from '@components/buttons/Button';
+import IconButton from '@components/buttons/IconButton';
+import TextButton from '@components/buttons/TextButton';
+import Layout from '@components/layout';
+import ArrowLink from '@components/links/ArrowLink';
+import ButtonLink from '@components/links/ButtonLink';
+import PrimaryLink from '@components/links/PrimaryLink';
+import UnderlineLink from '@components/links/UnderlineLink';
+import UnstyledLink from '@components/links/UnstyledLink';
+import NextImage from '@components/NextImage';
+import Seo from '@components/Seo';
+import Skeleton from '@components/Skeleton';
 
 type Color = (typeof colorList)[number];
