+++ src/pages/create-medical-record.tsx
@@ -1,8 +1,8 @@
-import DoctorProtectedRoute from '@/components/common/doctorProtectedRoute';
-import Container from '@/components/container';
-import CreateRecordForm from '@/components/doctors/createRecordForm';
-import Layout from '@/components/layout/Layout';
-import Seo from '@/components/Seo';
+import DoctorProtectedRoute from '@components/common/doctorProtectedRoute';
+import Container from '@components/container';
+import CreateRecordForm from '@components/doctors/createRecordForm';
+import Layout from '@components/layout';
+import Seo from '@components/Seo';
 
 function CreateMedicalRecordPage() {
   return (
