+++ src/pages/index.tsx
@@ -1,11 +1,11 @@
 import Link from 'next/link';
 import * as React from 'react';
 
-import Button from '@/components/buttons/Button';
-import { ProtectedRoute } from '@/components/common/protectedRoute';
-import Container from '@/components/container';
-import Layout from '@/components/layout/Layout';
-import Seo from '@/components/Seo';
+import Button from '@components/buttons/Button';
+import { ProtectedRoute } from '@components/common/protectedRoute';
+import Container from '@components/container';
+import Layout from '@components/layout';
+import Seo from '@components/Seo';
 
 function HomePage() {
   return (
