+++ src/pages/medical-records.tsx
@@ -2,13 +2,13 @@ import clsx from 'clsx';
 import { useState } from 'react';
 import { useQuery } from 'react-query';
 
-import { ProtectedRoute } from '@/components/common/protectedRoute';
-import Container from '@/components/container';
-import Layout from '@/components/layout/Layout';
-import MedicalRecordsList from '@/components/medicalRecordsList';
-import Seo from '@/components/Seo';
-import MedicalRecordsSkeleton from '@/components/skeletons/medicalRecords.skeleton';
-import MedicalRecordsPageSkeleton from '@/components/skeletons/medicalRecordsPage.skeleton';
+import { ProtectedRoute } from '@components/common/protectedRoute';
+import Container from '@components/container';
+import Layout from '@components/layout';
+import MedicalRecordsList from '@components/medicalRecordsList';
+import Seo from '@components/Seo';
+import MedicalRecordsSkeleton from '@components/skeletons/medicalRecords.skeleton';
+import MedicalRecordsPageSkeleton from '@components/skeletons/medicalRecordsPage.skeleton';
 
 import { getMedicalRecords } from '@/api/medicalRecords';
 import { MEDICAL_RECORDS_KEY } from '@/constant/queryKeys';
