+++ tsconfig.json
@@ -16,6 +16,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
+      "@components/*": ["./src/components/*"],
       "~/*": ["./public/*"]
     },
     "incremental": true
